elastic/pyproject.toml
[tool.poetry]
name = "elastic"
version = "0.1.0"
description = ""
authors = ["Arc1el <khm970514@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
uvicorn = "^0.20.0"
python-dotenv = "^0.21.1"
elasticsearch = "^8.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "elastic.main:start"